feat:热工预装炉查询，预装炉装配跟单查询等流程

+++ mes-common/common-model/src/main/java/com/richfit/mes/common/model/produce/PrechargeFurnace.java
@@ -19,7 +19,7 @@ public class PrechargeFurnace {
     private static final long serialVersionUID = -5801273490970600632L;
 
     @TableId(type = IdType.AUTO)
-    protected Integer id;
+    protected Long id;
 
     /**
      * 创建者
